requirements.txt
fastapi[all]
PyJWT
cffi
pyparser
uvicorn[standard]
rsa
sqlalchemy
mysql-connector-python
cryptography
google-auth
pytz
bcrypt
